fix(ios): centralize snapshot presentation ownership
