+++ src/core/__tests__/snapshot-state.test.ts
@@ -1,10 +1,8 @@
 import { expect, test } from 'vitest';
 import { buildSnapshotState } from '../snapshot-state.ts';
+import { resolveActionableTouchResolution } from '../interaction-targeting.ts';
 import { createSnapshotVisibility } from '@agent-device/contracts/snapshot';
-import {
-  attachSnapshotOcclusionContextEvidence,
-  attachSnapshotPresentationEvidence,
-} from '@agent-device/contracts/capture';
+import { attachSnapshotOcclusionContextEvidence } from '@agent-device/contracts/capture';
 import {
   buildUiHierarchySnapshot,
   parseUiHierarchyTree,
@@ -156,21 +154,18 @@ test('buildSnapshotState preserves the legacy acquired iOS presentation path', (
   ]);
 });
 
-test('buildSnapshotState skips the legacy path for engine-presented iOS results', () => {
+test('buildSnapshotState uses the registered presentation owner for Appium results', () => {
   const rowRect = { x: 16, y: 293, width: 370, height: 52 };
-  const data = attachSnapshotPresentationEvidence(
-    {
-      nodes: [
-        { index: 0, depth: 0, type: 'Application', label: 'Settings' },
-        { index: 1, depth: 1, parentIndex: 0, type: 'CollectionView' },
-        { index: 2, depth: 2, parentIndex: 1, type: 'Cell', label: 'General', rect: rowRect },
-        { index: 3, depth: 3, parentIndex: 2, type: 'Button', label: 'General', rect: rowRect },
-      ],
-      backend: 'xctest' as const,
-      producer: 'appium-source' as const,
-    },
-    { owner: 'ios-snapshot-engine' },
-  );
+  const data = {
+    nodes: [
+      { index: 0, depth: 0, type: 'Application', label: 'Settings' },
+      { index: 1, depth: 1, parentIndex: 0, type: 'CollectionView' },
+      { index: 2, depth: 2, parentIndex: 1, type: 'Cell', label: 'General', rect: rowRect },
+      { index: 3, depth: 3, parentIndex: 2, type: 'Button', label: 'General', rect: rowRect },
+    ],
+    backend: 'xctest' as const,
+    producer: 'appium-source' as const,
+  };
 
   const state = buildSnapshotState(data, { snapshotInteractiveOnly: true });
 
@@ -180,6 +175,49 @@ test('buildSnapshotState skips the legacy path for engine-presented iOS results'
     ['Cell', 'General'],
     ['Button', 'General'],
   ]);
+});
+
+test('Appium presentation does not infer hittability from an enabled ancestor rectangle', () => {
+  const state = buildSnapshotState(
+    {
+      nodes: [
+        {
+          index: 0,
+          depth: 0,
+          type: 'Application',
+          rect: { x: 0, y: 0, width: 390, height: 844 },
+        },
+        {
+          index: 1,
+          depth: 1,
+          parentIndex: 0,
+          type: 'Cell',
+          rect: { x: 16, y: 293, width: 370, height: 52 },
+          enabled: true,
+        },
+        {
+          index: 2,
+          depth: 2,
+          parentIndex: 1,
+          type: 'StaticText',
+          label: 'General',
+          rect: { x: 24, y: 309, width: 100, height: 20 },
+          enabled: true,
+        },
+      ],
+      backend: 'xctest',
+      producer: 'appium-source',
+    },
+    { snapshotInteractiveOnly: true },
+  );
+  const target = state.nodes.find((node) => node.label === 'General');
+
+  expect(target).toBeDefined();
+  expect(target?.hittable).toBeUndefined();
+  expect(resolveActionableTouchResolution(state.nodes, target!)).toMatchObject({
+    node: target,
+    reason: 'original',
+  });
 });
 
 test('buildSnapshotState marks content covered by floating overlays as visible but blocked', () => {
